UPDATE: CL-2719 Usage of direct memory by engine should be limited - for now only equivalent of half of heap size is used for direct clover buffers
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/branches/release-3-4-0-M2@13720 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/graph/runtime/jmx/GraphTracking.java
@@ -98,6 +98,6 @@ public interface GraphTracking extends Serializable {
 	/**
 	 * @return size of memory footprint for whole graph (not guaranteed)
 	 */
-	int getUsedMemory();
+	long getUsedMemory();
 
 }
